counterfactual_oracle/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
openai==1.3.0
google-generativeai==0.3.1
numpy==1.26.2
pandas==2.1.3
requests==2.31.0
fpdf==2.7.6
psycopg2-binary==2.9.9



